Improve refresh end animation
There’s actually no need to set the scroll view’s `contentInset` to
`UIEdgeInsetsZero` since f49413e because we now lift
`-setContentInsets:` with the `-contentInsets` signal. Now we’ll
animate closing the scroll view immediately upon the completion of the
refresh command. Once the view is closed, the insets will be updated to
their normal state.