Bind properties to output of updateWeatherCommand
We can use `executionSignals` and `errors` to receive the latest
successful or failure values from the signal that’s subscribed to via
the command. This lets us remove the `doNext` and `doError` side
effects from the signal itself. This also gives us a convenient place
to forward events or errors to analytics.