Implement custom paging
This paging algorithm is highly dependent on a velocity trend. If the
velocity trends positive, we prefer to page down. If the velocity
trends negative, we page up. If the dragging ends and the current
velocity is the same as the last velocity, we simply use whether the
velocity is currently positive or negative to decide whether to page
down or up respectively.