Cache weather updates to disk
This caches weather updates to disk
and loads them first when the app boots up.

This provides a nicer experience for the user because
they don't see a blank screen every time they open the app
because it's waiting to check the weather.

This also prepares us to impement background fetching
because all we need to do is fetch an update and store it in the cache
for it to be loaded first for the user.

- Create an ApplicationController and instantiate it in the AppDelegate
- Implement NSCoding on TRWeatherUpdate
- Add TRWeatherUpdateCache to cache TRWeatherUpdates
- Load the weather from the cache on boot