Fix test failure when running in PST
Previously the current calendar's time zone was configured to UTC, but
when a date was constructed using `NSDateFormatter` the default time
zone would be used. When the default time zone is PST, the created time
fell into the "afternoon" range instead of the expected "night" range,
and thus the test would fail.

Instead of updating the current calendar's timezone, we instead update
the default timezone, so that the date formatter is in the correct time
zone when the test runs.