Don't perform background fetch when receiving push notification
When the app was launched in the background in response to a silent
notification, no local notification was ever scheduled.

Inspecting the logs revealed that an error prevented the weather
description from being updated because the `updateWeatherCommand` is
already executing. This is because when the app is coming from suspended
state, `application:didFinishLaunchingWithOptions:` is called, which
triggers the `updateWeatherCommand`.

The solution is to not trigger the `updateWeatherCommand` when receiving
the notification. When the app is still active, the weather is very
likely to already be up-to-date. If it isn't active the weather will be
updated because `application:didFinishLaunchingWithOptions:` is called.