Fix integer parsing in Swift 4
In either Swift 3 or 4, implicit numeric casting was removed when
downcasting from `Any`. That is, previously the underlying `Double`
value was automatically converted to `Int` in the bridging cast, but now
the same cast returns `nil` instead.

Roundtripping via `NSNumber` explicitly allows us to restore the
original behaviour of the cast.