Add bridged Swift version of ForecastController
This change removes the forward-declaration of `TRWeatherUpdate` from
`TRForecastController.h`, because depending on this header from Swift,
which in turn depends on a class exported to Objective-C from Swift,
creates a cycle that causes the method to not be exported back to Swift.

Removing the forward-declaration and force-casting the method return
value to `WeatherUpdate` dynamically breaks the cycle while retaining
type safety on the Swift side.