Fix crash when running application-hosted tests
The TroposTests target was crashing on CI, but the stack trace from the
crash was being suppressed by `xcpretty`. The cause of the crash was the
invalid secrets triggering a parameter assert in `TRForecastController`.
Moving the assertions to the app delegate means that they are
effectively skipped during testing, allowing CI to continue to operate
without valid secrets.